requirements.txt
fastapi
uvicorn
opencv-python-headless
ultralytics
pyserial
numpy